codeql caught this

+++ webcam.cpp
@@ -57,7 +57,7 @@ image_t *webcam_read() {
 
   // Copy OpenCV Mat data directly to image_t
   // OpenCV Mat stores data as contiguous RGB bytes
-  const size_t data_size = frame.cols * frame.rows * 3; // 3 bytes per RGB pixel
+  const size_t data_size = (size_t)frame.cols * frame.rows * 3; // 3 bytes per RGB pixel
 
   memcpy(img->pixels, frame.data, data_size);
